LayoutTests/imported/w3c/web-platform-tests/push-api/idlharness.https.any-expected.txt


PASS idl_test setup
PASS idl_test validation
PASS Partial interface ServiceWorkerRegistration: original interface defined
PASS Partial interface ServiceWorkerRegistration: member names are unique
PASS Partial interface ServiceWorkerGlobalScope: original interface defined
PASS Partial interface ServiceWorkerGlobalScope: valid exposure set
PASS Partial interface ServiceWorkerGlobalScope: member names are unique
PASS WorkerGlobalScope includes WindowOrWorkerGlobalScope: member names are unique
FAIL PushManager interface: existence and properties of interface object assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface object length assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface object name assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface: existence and properties of interface prototype object assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface: attribute supportedContentEncodings assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface: operation subscribe(optional PushSubscriptionOptionsInit) assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface: operation getSubscription() assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface: operation permissionState(optional PushSubscriptionOptionsInit) assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager must be primary interface of registration.pushManager assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of registration.pushManager assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL PushManager interface: registration.pushManager must inherit property "supportedContentEncodings" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL PushManager interface: registration.pushManager must inherit property "subscribe(optional PushSubscriptionOptionsInit)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL PushManager interface: calling subscribe(optional PushSubscriptionOptionsInit) on registration.pushManager with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL PushManager interface: registration.pushManager must inherit property "getSubscription()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL PushManager interface: registration.pushManager must inherit property "permissionState(optional PushSubscriptionOptionsInit)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL PushManager interface: calling permissionState(optional PushSubscriptionOptionsInit) on registration.pushManager with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
PASS PushSubscriptionOptions interface: existence and properties of interface object
PASS PushSubscriptionOptions interface object length
PASS PushSubscriptionOptions interface object name
PASS PushSubscriptionOptions interface: existence and properties of interface prototype object
PASS PushSubscriptionOptions interface: existence and properties of interface prototype object's "constructor" property
PASS PushSubscriptionOptions interface: existence and properties of interface prototype object's @@unscopables property
PASS PushSubscriptionOptions interface: attribute userVisibleOnly
PASS PushSubscriptionOptions interface: attribute applicationServerKey
PASS PushSubscription interface: existence and properties of interface object
PASS PushSubscription interface object length
PASS PushSubscription interface object name
PASS PushSubscription interface: existence and properties of interface prototype object
PASS PushSubscription interface: existence and properties of interface prototype object's "constructor" property
PASS PushSubscription interface: existence and properties of interface prototype object's @@unscopables property
PASS PushSubscription interface: attribute endpoint
PASS PushSubscription interface: attribute expirationTime
PASS PushSubscription interface: attribute options
PASS PushSubscription interface: operation getKey(PushEncryptionKeyName)
PASS PushSubscription interface: operation unsubscribe()
PASS PushSubscription interface: operation toJSON()
PASS PushMessageData interface: existence and properties of interface object
PASS PushEvent interface: existence and properties of interface object
PASS PushSubscriptionChangeEvent interface: existence and properties of interface object
FAIL ServiceWorkerRegistration interface: attribute pushManager assert_true: The prototype object must have a property "pushManager" expected true got false
FAIL ServiceWorkerRegistration interface: registration must inherit property "pushManager" with the proper type assert_inherits: property "pushManager" not found in prototype chain
